requirements.txt
# requirements.txt
opcua
psycopg2-binary
PyQt5
